+++ src/Infocom.php
@@ -719,6 +719,7 @@ class Infocom extends CommonDBChild
 
                           $data['warrantyexpiration']        = $warranty;
                           $data['item_name']                 = $item_infocom->getName();
+                          $data['is_deleted']                = $item_infocom->maybeDeleted() ? (int) $item_infocom->fields['is_deleted'] : 0;
                           $items_infos[$entity][$data['id']] = $data;
 
                         if (!isset($items_messages[$entity])) {
@@ -731,9 +732,14 @@ class Infocom extends CommonDBChild
         }
 
         foreach ($items_infos as $entity => $items) {
+            // We will ignore items that have been deleted, but we won't add an alert to the DB in case they are restored before the warranty expires
+            $not_deleted_items = array_filter($items, static function ($item) {
+                return $item['is_deleted'] === 0;
+            });
             if (
-                NotificationEvent::raiseEvent("alert", new self(), ['entities_id' => $entity,
-                    'items'       => $items
+                NotificationEvent::raiseEvent("alert", new self(), [
+                    'entities_id' => $entity,
+                    'items'       => $not_deleted_items
                 ])
             ) {
                 $message     = $items_messages[$entity];
@@ -757,9 +763,11 @@ class Infocom extends CommonDBChild
                 }
 
                 $alert             = new Alert();
-                $input["itemtype"] = 'Infocom';
-                $input["type"]     = Alert::END;
-                foreach ($items as $id => $item) {
+                $input = [
+                    'itemtype' => 'Infocom',
+                    'type'     => Alert::END
+                ];
+                foreach ($not_deleted_items as $id => $item) {
                     $input["items_id"] = $id;
                     $alert->add($input);
                     unset($alert->fields['id']);